Evaluation with environments and definitions with define and let
Delayed committing this until I was confident I wanted to implement a
new data structure to represent the abstract syntax tree. I decided to
go with a basic environment only consisting of a map binding the
symbols to an associated expression. This way we can store functions,
symbols, numbers, etc alike all in the same `std::map`.

Currently this requires constructing a new vector every to evaluate a
function. This is used to represent the arguments to a function. A
better option to consider is passing an iterator since this will
remove the need to constantly construct a new vector for every set of
arguments.

In addition, this current implementation lacks the ability for let
blocks to access outer scopes. I still have yet to determine the best
plan of action here.

Garbage collection has not been touched yet but considering closures
will need them I need to consider how I can use shared pointers to
track referenced bindings and environments.